refactor: move headers folder under ui folder
